Rename Machine -> createMachine and .nodes -> ._nodes
