+++ packages/core/src/index.ts
@@ -8,7 +8,7 @@ export * from './typegenTypes.ts';
 export * from './types.ts';
 export { waitFor } from './waitFor.ts';
 import { Actor, createActor, interpret, Interpreter } from './interpreter.ts';
-import { createMachine } from './Machine.ts';
+import { createMachine } from './createMachine.ts';
 export { type MachineSnapshot, isMachineSnapshot } from './State.ts';
 import { StateNode } from './StateNode.ts';
 // TODO: decide from where those should be exported
