+++ packages/core/test/final.test.ts
@@ -848,7 +848,7 @@ describe('final states', () => {
     const actorRef = createActor(machine).start();
     flushTracked();
 
-    // it's important to send an event here that results in a transition that computes new `state.nodes`
+    // it's important to send an event here that results in a transition that computes new `state._nodes`
     // and that could impact the order in which exit actions are called
     actorRef.send({ type: 'EV' });
 
@@ -899,7 +899,7 @@ describe('final states', () => {
 
     const actorRef = createActor(machine).start();
 
-    // it's important to send an event here that results in a transition as that computes new `state.nodes`
+    // it's important to send an event here that results in a transition as that computes new `state._nodes`
     // and that could impact the order in which exit actions are called
     actorRef.send({ type: 'EV1' });
     flushTracked();
@@ -954,7 +954,7 @@ describe('final states', () => {
     const flushTracked = trackEntries(machine);
 
     const actorRef = createActor(machine).start();
-    // it's important to send an event here that results in a transition as that computes new `state.nodes`
+    // it's important to send an event here that results in a transition as that computes new `state._nodes`
     // and that could impact the order in which exit actions are called
     actorRef.send({ type: 'EV1' });
     flushTracked();
@@ -1010,7 +1010,7 @@ describe('final states', () => {
 
     const actorRef = createActor(machine).start();
     flushTracked();
-    // it's important to send an event here that results in a transition as that computes new `state.nodes`
+    // it's important to send an event here that results in a transition as that computes new `state._nodes`
     // and that could impact the order in which exit actions are called
     actorRef.send({ type: 'EV' });
