+++ packages/core/test/input.test.ts
@@ -1,6 +1,6 @@
 import { of } from 'rxjs';
 import { assign, createActor, spawn } from '../src';
-import { createMachine } from '../src/Machine';
+import { createMachine } from '../src/createMachine';
 import {
   fromCallback,
   fromObservable,
